refactor(ts): refactor
