+++ src/typescript/src/index.ts
@@ -1,6 +1,7 @@
 import yargs from 'yargs';
 import Crawler from "./core/crawler";
 import Argv from "./interfaces/argv";
+import { loadDotEnv } from "./utils/laod-dot-env";
 
 yargs.command('crawl <spiderName>', 'run the spider', (yargs) => {
     yargs.positional('name', {
@@ -31,5 +32,6 @@ yargs.command('crawl <spiderName>', 'run the spider', (yargs) => {
 
 
 async function main(argv: Argv) {
+    loadDotEnv();
     await Crawler.run(argv);
 }
