+++ src/typescript/src/utils/logger.ts
@@ -1,31 +1,42 @@
 import winston from "winston";
-import { Logger as LoggerInterface, format } from "winston";
+import { Logger as LoggerInterface, format, level } from "winston";
 
 export enum levels {
-    DEBUG,
-    INFO,
-    WARN,
-    // TODO: WARNING
-    ERROR,
-    CRITICAL
+    ERROR = 0,
+    WARN = 1,
+    INFO = 2,
+    HTTP = 3,
+    VERBOSE = 4,
+    DEBUG = 5,
+    SILLY = 6
 }
 
 export class Logger {
     public static createLogger(name: string, level: levels): LoggerInterface {
         const stringLevel = levels[level].toLowerCase();
 
-        const consoleTransport = new winston.transports.Console({
-            level: stringLevel,
-            format: winston.format.combine(
-                winston.format.colorize(),
-                winston.format.timestamp({ format: "YYYY-MM-DD HH:mm:ss" }),
-                this.scrapyLogFormat()
-            )
-        });
+        const transport = !!process.env.LOG_FILE && process.env.LOG_FILE.length
+            ? new winston.transports.File({
+                // options: { flags: 'a', encoding: 'utf8' },
+                filename: 'log.log',
+                level: stringLevel,
+                format: winston.format.combine(
+                    winston.format.timestamp({ format: "YYYY-MM-DD HH:mm:ss" }),
+                    this.scrapyLogFormat()
+                )
+            })
+            : new winston.transports.Console({
+                level: stringLevel,
+                format: winston.format.combine(
+                    winston.format.colorize(),
+                    winston.format.timestamp({ format: "YYYY-MM-DD HH:mm:ss" }),
+                    this.scrapyLogFormat()
+                )
+            });
 
         const logger = winston.createLogger({
             level: stringLevel,
-            transports: [consoleTransport],
+            transports: [transport],
         });
 
         return logger.child({
@@ -42,7 +53,7 @@ export class Logger {
 
     private static scrapyLogFormat() {
         return format.printf(({ level, message, label, timestamp }) => {
-            return `${timestamp} [${label}] ${level.toUpperCase()}: ${message}`
+            return `${timestamp} [${label}] ${level.trim().toUpperCase()}: ${message}`;
         });
     }
 }
